requirements.txt
numpy
soundfile
torch
torchaudio
transformers>=4.57.0,<4.58.0
accelerate
huggingface_hub
diffusers
einops
loguru
tqdm
numba>=0.63.1
vector-quantize-pytorch
PyYAML
modelscope
filelock>=3.13.0
peft>=0.11.0
gradio>=4.0.0
pandas
bitsandbytes
fastapi
uvicorn
python-multipart
openai
python-dotenv
